Revert "Move delay/retry function to retryLink file"
This reverts commit b92973cbcbe7b14d404256f0f1ca7c99e4da3f95.
