+++ src/link/retry/retryLink.ts
@@ -8,6 +8,9 @@ import {
 } from "@apollo/client/errors";
 import { ApolloLink } from "@apollo/client/link";
 
+import { buildDelayFunction } from "./delayFunction.js";
+import { buildRetryFunction } from "./retryFunction.js";
+
 export declare namespace RetryLink {
   export type DelayFunction = (
     count: number,
@@ -199,37 +202,4 @@ export class RetryLink extends ApolloLink {
       };
     });
   }
-}
-
-function buildDelayFunction(
-  delayOptions?: RetryLink.DelayOptions
-): RetryLink.DelayFunction {
-  const { initial = 300, jitter = true, max = Infinity } = delayOptions || {};
-  // If we're jittering, baseDelay is half of the maximum delay for that
-  // attempt (and is, on average, the delay we will encounter).
-  // If we're not jittering, adjust baseDelay so that the first attempt
-  // lines up with initialDelay, for everyone's sanity.
-  const baseDelay = jitter ? initial : initial / 2;
-
-  return function delayFunction(count: number) {
-    let delay = Math.min(max, baseDelay * 2 ** count);
-    if (jitter) {
-      // We opt for a full jitter approach for a mostly uniform distribution,
-      // but bound it within initialDelay and delay for everyone's sanity.
-      delay = Math.random() * delay;
-    }
-
-    return delay;
-  };
-}
-
-function buildRetryFunction(
-  retryOptions?: RetryLink.AttemptsOptions
-): RetryLink.AttemptsFunction {
-  const { retryIf, max = 5 } =
-    retryOptions || ({} as RetryLink.AttemptsOptions);
-  return function retryFunction(count, operation, error) {
-    if (count >= max) return false;
-    return retryIf ? retryIf(error, operation) : !!error;
-  };
 }
